Move skin code to a separate module
mesh.rs is infamously large. We could split off unrelated code.

Morph targets are very similar to skinning and have their own module.
We move skinned meshes to an independent module like morph targets and
give the systems similar names.
