+++ crates/bevy_pbr/src/render/mod.rs
@@ -3,8 +3,10 @@ mod light;
 pub(crate) mod mesh;
 mod mesh_bindings;
 mod morph;
+mod skin;
 
 pub use fog::*;
 pub use light::*;
 pub use mesh::*;
 pub use mesh_bindings::MeshLayouts;
+pub use skin::{extract_skins, prepare_skins, SkinIndex, SkinUniform, MAX_JOINTS};
